+++ lib/format.dart
@@ -1,7 +1,16 @@
 export 'src/double_format.dart'
     show DoubleFormatMode, DoubleSpecialValueSpelling;
 export 'src/engine.dart'
-    show Format, defaultFormat, format, formatWith, sprintf, vsprintf;
+    show
+        Format,
+        clearTemplateCache,
+        defaultFormat,
+        format,
+        formatWith,
+        sprintf,
+        templateCacheCapacity,
+        templateCacheSize,
+        vsprintf;
 export 'src/errors.dart'
     show
         AmbiguousFormatterException,
@@ -18,4 +27,4 @@ export 'src/errors.dart'
 export 'src/extensions.dart'
     show AttributeLookup, FormatOptions, Formatter, Representation;
 export 'src/number_locale.dart' show CNumberLocale, NumberLocale;
-export 'src/text_unit.dart' show TextUnit;
+export 'src/text_unit.dart' show TextUnit, TextUnitOperations;